Initial version
Performs the OAuth2 negotation and gets a token, uses a few public endpoints to gather more detailed information about where the authed character comes from and can log some of the information to the console.  Stores all of this in a temporary server side session, next edit will integrate the upcoming auth-srv.